+++ packages/scene/src/index.ts
@@ -106,6 +106,7 @@ class Scene
 
   public addLayer(layer: ILayer): void {
     // 为当前图层创建一个容器
+    // TODO: 初始化的时候设置 容器
     const layerContainer = createLayerContainer(this.container);
     layer.setContainer(layerContainer);
     this.sceneService.addLayer(layer);
@@ -210,6 +211,9 @@ class Scene
     this.mapService.panTo(pixel);
   }
 
+  public setZoom(zoom: number): void {
+    this.mapService.setZoom(zoom);
+  }
   public fitBounds(bound: Bounds): void {
     this.mapService.fitBounds(bound);
   }